pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ufc-lakehouse"
version = "0.1.0"
description = "Lakehouse-style pipeline (RAW/Bronze/Silver/Gold) para UFC Stats"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "PyYAML>=6.0",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
parquet = [
  "pandas>=2.1",
  "pyarrow>=14.0",
]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
  "mypy>=1.7",
  "types-PyYAML",
  "types-requests",
]

[project.scripts]
ufc-pipeline = "ufc_pipeline.orchestration.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
